requirements.txt
python-dotenv
gradio
openai
pydub